# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CSV Import & Report API",
        "description": "Upload order CSVs, track asynchronous import jobs, and request asynchronous reports.\n\n**Authentication (demo):** every request identifies the caller with the `X-User-Id` header (click *Authorize* and enter e.g. `alice`). Users only see their own imports and reports.\n\n**Idempotency:** POST endpoints accept an optional `Idempotency-Key` header; repeating a request with the same key returns the original job instead of creating a new one.",
        "license": {
            "name": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/imports": {
            "get": {
                "tags": [
                    "imports"
                ],
                "summary": "List your imports, newest first.",
                "operationId": "list_imports",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Page size (1-500, default 50).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Items to skip (default 0).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportList"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "user_id": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "imports"
                ],
                "summary": "Upload a CSV file and start an asynchronous import.",
                "description": "Returns `202 Accepted` immediately with the import id; poll `GET /imports/{id}` for progress.",
                "operationId": "create_import",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "Optional key to make retries of this request safe",
                        "required": false,
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UploadForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Import job created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportAccepted"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No file / empty file",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid X-User-Id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "user_id": []
                    }
                ]
            }
        },
        "/imports/{id}": {
            "get": {
                "tags": [
                    "imports"
                ],
                "summary": "Get the status and progress of an import.",
                "operationId": "get_import",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Import id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Import status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportStatus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found (or belongs to another user)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "user_id": []
                    }
                ]
            }
        },
        "/imports/{id}/errors": {
            "get": {
                "tags": [
                    "imports"
                ],
                "summary": "List the invalid rows of an import and why each was rejected.",
                "operationId": "list_import_errors",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Import id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Page size (1-500, default 50).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Items to skip (default 0).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invalid rows",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportErrorsPage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "user_id": []
                    }
                ]
            }
        },
        "/imports/{id}/file": {
            "get": {
                "tags": [
                    "imports"
                ],
                "summary": "Download the originally uploaded CSV file from object storage.",
                "operationId": "download_import_file",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Import id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The uploaded file",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "The file was removed with storage-admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Object storage unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorBody"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "user_id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorBody": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                }
            },
...
            "ImportAccepted": {
                "type": "object",
                "required": [
                    "import_id",
                    "status",
                    "links"
                ],
                "properties": {
                    "import_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "links": {
                        "$ref": "#/components/schemas/Links"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ImportState"
                    }
                }
            },
...
            "ImportList": {
                "type": "object",
                "required": [
                    "items",
                    "limit",
                    "offset"
                ],
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImportSummary"
                        }
                    },
                    "limit": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ImportState": {
                "type": "string",
                "description": "User-facing import state.",
                "enum": [
                    "queued",
                    "retrying",
                    "processing",
                    "completed",
                    "completed_with_errors",
                    "failed"
                ]
            },
            "ImportStatus": {
                "type": "object",
                "required": [
                    "import_id",
                    "filename",
                    "file_size",
                    "file_sha256",
                    "file_available",
                    "status",
                    "attempts",
                    "max_attempts",
                    "created_at",
                    "attempt_history",
                    "links"
                ],
                "properties": {
                    "attempt_history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttemptInfo"
                        }
                    },
                    "attempts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "file_available": {
                        "type": "boolean",
                        "description": "Whether the uploaded file is still in object storage (see `GET /imports/{id}/file`)."
                    },
                    "file_deleted_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time",
                        "description": "When the file was removed with `storage-admin` (imported rows are kept)."
                    },
                    "file_sha256": {
                        "type": "string"
                    },
                    "file_size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "finished_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "import_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "invalid_rows": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32"
                    },
                    "last_error": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Error of the most recent failed attempt (or the final failure reason)."
                    },
                    "links": {
                        "$ref": "#/components/schemas/Links"
                    },
                    "max_attempts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "next_attempt_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time",
                        "description": "When the next attempt is scheduled (only while `retrying`)."
                    },
                    "started_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ImportState"
                    },
                    "total_rows": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32",
                        "description": "Data rows in the file (excluding the header). Null until processed."
                    },
                    "valid_rows": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32"
                    }
                }
            },
...
            "UploadForm": {
                "type": "object",
                "description": "Multipart form for uploads.",
                "required": [
                    "file"
                ],
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "description": "The CSV file. Required columns: user_id,order_id,product,quantity,unit_price,status"
                    }
                }
            }
        },
        "securitySchemes": {
            "user_id": {
                "type": "apiKey",
                "in": "header",
                "name": "X-User-Id",
                "description": "Caller identity (demo auth). Letters, digits, '-', '_', '.', '@'; max 64 chars."
            }
        }
    },
...
}